Move InvocationHandlerFactory to internal
